Refactoring code for consistency
Making all providers mimic the terraform provider by separating the config
definition into its own file.
